requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
elevenlabs==0.2.24
mutagen==1.47.0
eyed3==0.9.7
musicbrainzngs==0.7.1
pydub==0.25.1
spotipy==2.23.0
pylast==5.2.0
praw==7.7.1
schedule==1.2.0
python-m3u8==3.4.0
numpy==1.24.3
pandas==2.0.3
